integrating tour package with the current location of logged user

+++ app/src/main/java/com/example/tripto/retrofit/ApiInterface.kt
@@ -68,7 +68,7 @@ interface ApiInterface {
         @GET("/getRatedPlaces")
         fun getRatedPlaces(@Query("user_id") user_id: Int): Call<List<getRatedPlacesModel>>
         @GET("/getTourPackage")
-        fun getTourPackage(@Query("user_id") user_id: Int,@Query("longitude") longitude: Int,@Query("latitude") latitude: Int): Call<List<PlaceModel>>
+        fun getTourPackage(@Query("user_id") user_id: Int,@Query("longitude") longitude: Double,@Query("latitude") latitude: Double): Call<List<PlaceModel>>
         companion object {
         fun create(): ApiInterface {
             val retrofit = Retrofit.Builder()
